Some more comments addressed
