+++ node/src/integration-test/kotlin/net/corda/node/services/BFTNotaryServiceTests.kt
@@ -25,7 +25,7 @@ import net.corda.node.services.transactions.minClusterSize
 import net.corda.node.services.transactions.minCorrectReplicas
 import net.corda.nodeapi.internal.DevIdentityGenerator
 import net.corda.nodeapi.internal.network.NetworkParametersCopier
-import net.corda.core.internal.NotaryInfo
+import net.corda.core.node.NotaryInfo
 import net.corda.testing.core.chooseIdentity
 import net.corda.testing.common.internal.testNetworkParameters
 import net.corda.testing.contracts.DummyContract
